refactor(direct-control): extract unit-target action types
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only unit-target action input/candidate/result types into packages/civ7-direct-control/src/play/operations/unit-target-action.ts. Keeps public facade type re-exports in index.ts and leaves broader generic operation type ownership pending.

Preserves public type contracts and does not change unit-target source strings, command serialization, parser labels, wrapper ownership, approval-first behavior, post-send verification timing/wording, no-repeat-after-unverified guidance, runtime proof status, telemetry, AI ingestion, hotseat runtime proof, CLI semantic projection, Effect/oRPC procedure-core work, or Task 2.9.4 matrix-row acceptance. Generic operation, diplomacy/narrative closeout, public procedure schema, and telemetry type ownership remain pending separate owner slices.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- unit-target-action.test.ts public-api.test.ts
- bun run --cwd packages/cli test -- game.play.unit-target.test.ts
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- unit-target type ownership scan and unit-target no-cycle import scan

Local package/type relocation proof only; no runtime/live-game, unit-target behavior, CLI semantic projection, telemetry, AI-ingestion, hotseat product-path, or procedure-core proof claimed.
