+++ openspec/changes/civ7-support-direct-control-modularization/tasks.md
@@ -470,6 +470,13 @@ runtime/direct-control claims.
         operation/proof telemetry, hotseat runtime proof, AI ingestion, CLI
         semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4
         matrix-row acceptance pending.
+  - [x] 4.8.24 Extract unit-target action type owner while keeping public
+        facade type re-exports in `index.ts`, preserving unit-target
+        input/candidate/result contracts while leaving generic operation,
+        diplomacy/narrative closeout, public procedure schemas,
+        operation/proof telemetry, hotseat runtime proof, AI ingestion, CLI
+        semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4
+        matrix-row acceptance pending.
 - [ ] 4.9 Import or explicitly cite oRPC architecture authority before any
       oRPC implementation; current support branch has no tracked
       `.agents/skills/civ7-orpc-control-architecture` or
