+++ openspec/changes/civ7-support-direct-control-modularization/workstream/workstream-record.md
@@ -779,6 +779,19 @@ All future agent waves must be framed before delegation:
   procedure schema, and telemetry type ownership remain pending separate owner
   slices. This is local package/type relocation proof only, not runtime proof,
   AI/hotseat product-path support, or procedure-core readiness.
+- Direct-control unit-target action types slice: completed as a narrow public
+  type ownership relocation. It moves only unit-target action input/candidate/
+  result types into `src/play/operations/unit-target-action.ts` while keeping
+  facade type re-exports in `index.ts`. This preserves public type contracts
+  and does not change unit-target source strings, command serialization, parser
+  labels, wrapper ownership, approval-first behavior, post-send verification
+  timing/wording, no-repeat-after-unverified guidance, runtime proof status,
+  telemetry, AI ingestion, hotseat runtime proof, CLI semantic projection,
+  Effect/oRPC procedure-core work, or Task 2.9.4 matrix-row acceptance. Generic
+  operation, diplomacy/narrative closeout, public procedure schema, and
+  telemetry type ownership remain pending separate owner slices. This is local
+  package/type relocation proof only, not runtime proof, AI/hotseat product-path
+  support, or procedure-core readiness.
 - Direct-control turn-completion slice: completed as a narrow turn-completion
   source/wrapper relocation. It moves only `getCiv7TurnCompletionStatus`,
   `sendCiv7TurnComplete`, `sendCiv7TurnUnready`, the status command builder,
